+++ lib/distribute_reads/global_methods.rb
@@ -23,7 +23,7 @@ module DistributeReads
             current_lag =
               begin
                 DistributeReads.replication_lag(connection: base_model.connection)
-              rescue DistributeReads::NoReplicasAvailable
+              rescue ActiveRecord::ConnectionNotEstablished
                 # TODO rescue more exceptions?
                 false
               end